Course

+++ Course Management Backend/courses/models/course.py
@@ -7,8 +7,8 @@ class Course(BaseModel):
     title = models.CharField(max_length=200)
     description = models.TextField()
     banner = models.ImageField(upload_to="Course/" , blank=True , null=True , default="Course/Default_Course.jpg")
-    price = models.DecimalField(max_digits=10, decimal_places=2)
-    duration = models.DecimalField(max_digits=5 , decimal_places=2)
+    price = models.IntegerField()
+    duration = models.IntegerField()
     category = models.ForeignKey(Category, on_delete=models.CASCADE)
